Remove Deprecated OpenSAML 3 Support
Closes gh-10556
