+++ gradle.properties
@@ -2,7 +2,7 @@ aspectjVersion=1.9.9.1
 springJavaformatVersion=0.0.34
 springBootVersion=2.4.2
 springFrameworkVersion=6.0.0-M6
-openSamlVersion=3.4.6
+openSamlVersion=4.1.1
 version=6.0.0-SNAPSHOT
 kotlinVersion=1.7.10
 samplesBranch=main
